Project/Line Bot - Weather/requirements.txt
firebase_admin==6.5.0
line_bot_sdk==3.11.0
Requests==2.31.0
